Joomla 3.9+ compatibility and Ko-fi donation button
Added a patching system to patch the plugin when it is installed on Joomla 3.9+ to make it work. The patch replaces some classes in the manifest xml file and renames the language files to make them readable by Joomla 3.

Added Ko-fi donation button in the donation window.

Added GitHub links in the support window.

Replaced the debug window title.

Minor changes in en-GB main text file.